--- TMDT_Banmypham/Controllers/HomePageController.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-
-namespace TMDT_Banmypham.Controllers
-{
-    public class HomePageController : Controller
-    {
-        public IActionResult Index()
-        {
-            return View();
-        }
-    }
-}
