+++ TMDT_Banmypham/Models/DB/Baiviet.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace TMDT_Banmypham.Models.DB
+{
+    public partial class Baiviet
+    {
+        public string MaBv { get; set; }
+        public string TieuDeBv { get; set; }
+        public string MoTaBv { get; set; }
+    }
+}
