--- TMDT_Banmypham/Models/Hinhanh.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace TMDT_Banmypham.Models
-{
-    public partial class Hinhanh
-    {
-        public string MaHa { get; set; }
-        public string TenHa { get; set; }
-        public string Link { get; set; }
-    }
-}
