+++ TMDT_Banmypham/Views/Home/SanPham.cshtml
@@ -3,7 +3,9 @@
     ViewData["Title"] = "SanPham";
     Layout = "~/Views/Shared/_Layout.cshtml";
 }
-
+<head>
+    <link href="~/css/style_Baiviet.css" rel="stylesheet" />""
+</head>
 
 <!--<nav class="jss4 jss5 navbar navbar-expand-md navbar-light">
     <div class="container">
