+++ TMDT_Banmypham/Views/Sanphams/Edit.cshtml
@@ -1,0 +1,57 @@
+﻿@model TMDT_Banmypham.Models.DB.Sanpham
+
+@{
+    ViewData["Title"] = "Edit";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<h1>Edit</h1>
+
+<h4>Sanpham</h4>
+<hr />
+<div class="row">
+    <div class="col-md-4">
+        <form asp-action="Edit">
+            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+            <input type="hidden" asp-for="MaSp" />
+            <div class="form-group">
+                <label asp-for="TenSp" class="control-label"></label>
+                <input asp-for="TenSp" class="form-control" />
+                <span asp-validation-for="TenSp" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="ThuongHieu" class="control-label"></label>
+                <input asp-for="ThuongHieu" class="form-control" />
+                <span asp-validation-for="ThuongHieu" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="GiaSp" class="control-label"></label>
+                <input asp-for="GiaSp" class="form-control" />
+                <span asp-validation-for="GiaSp" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="MoTaSp" class="control-label"></label>
+                <input asp-for="MoTaSp" class="form-control" />
+                <span asp-validation-for="MoTaSp" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="Hinhanh" class="control-label"></label>
+                <input asp-for="Hinhanh" class="form-control" />
+                <span asp-validation-for="Hinhanh" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="MaLsp" class="control-label"></label>
+                <select asp-for="MaLsp" class="form-control" asp-items="ViewBag.MaLsp"></select>
+                <span asp-validation-for="MaLsp" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <input type="submit" value="Save" class="btn btn-primary" />
+            </div>
+        </form>
+    </div>
+</div>
+
+<div>
+    <a asp-action="Index">Back to List</a>
+</div>
+
