+++ TMDT_Banmypham/Views/Shared/_Layout.cshtml
@@ -23,6 +23,7 @@
     <link rel="stylesheet" href="~/css/dang ky.css" />
     <link rel="stylesheet" href="~/css/dang nhap.css" />
     <link rel="stylesheet" href="~/css/darkmode.css" />
+    
     <!--Jquery-->
     <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"
             integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj"
