+++ TMDT_Banmypham/appsettings.json
@@ -1,6 +1,7 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=aspnet-TMDT_Banmypham-FC2AF320-950C-418A-8089-F49C05E9760F;Trusted_Connection=True;MultipleActiveResultSets=true"
+    "DefaultConnection": "Server=DESKTOP-K7V6G3D\\SQLEXPRESS;Database=aspnet-TMDT_Banmypham-FC2AF320-950C-418A-8089-F49C05E9760F;Trusted_Connection=True;MultipleActiveResultSets=true",
+    "MYPHAMdatabase": "Server=DESKTOP-K7V6G3D\\SQLEXPRESS;Database=MYPHAM;Trusted_Connection=True;"
   },
   "Logging": {
     "LogLevel": {
